Refactor out LK specific code from ProxyServlet
